DAYD-734 delete guestCheckを利用しない実装に変更
