+++ src/components/utils/RedirectToMainIfLoginComponent.tsx
@@ -1,4 +1,4 @@
-import { useGuestCheckQuery } from '@/redux/auth/slice';
+import { useReadUserQuery } from '@/redux/user/slice';
 import { useRouter } from 'next/router';
 
 type Props = {
@@ -7,10 +7,10 @@ type Props = {
 
 export const RedirectToMainIfLoginComponent = (props: Props) => {
   const router = useRouter();
-  const { data: guestCheckResult, isFetching } = useGuestCheckQuery();
+  const { data: User, isFetching, isError } = useReadUserQuery();
 
-  if (!isFetching && guestCheckResult) {
-    if (!guestCheckResult.isLogin) {
+  if (!isFetching && User) {
+    if (isError) {
       return props.children;
     } else {
       router.replace('/main');
